make acme optional
