+++ Cargo.toml
@@ -8,7 +8,9 @@ repository = "https://github.com/dfinity/ic-bn-lib"
 readme = "README.md"
 
 [features]
-clients-hyper = ["hyper-rustls"]
+clients-hyper = ["dep:hyper-rustls"]
+acme_dns = ["dep:instant-acme", "dep:cloudflare", "dep:backoff", "dep:rcgen"]
+acme_alpn = ["dep:rustls-acme"]
 
 [dependencies]
 ahash = "0.8.11"
@@ -16,14 +18,14 @@ anyhow = "1.0.93"
 arc-swap = "1.7.1"
 async-trait = "0.1.83"
 axum = "0.8.1"
-backoff = { version = "0.4.0", features = ["tokio"] }
+backoff = { version = "0.4.0", features = ["tokio"], optional = true }
 base64 = "0.22.1"
 bytes = "1.10.0"
 clap = { version = "4.5.20", features = ["derive", "string", "env"] }
 chacha20poly1305 = "0.10.1"
 cloudflare = { version = "0.13.0", default-features = false, features = [
     "rustls-tls",
-] }
+], optional = true }
 derive-new = "0.7.0"
 fqdn = "0.4.1"
 futures = "0.3.31"
@@ -53,14 +55,14 @@ hyper-util = { version = "0.1.10", features = ["full"] }
 instant-acme = { version = "0.7.2", default-features = false, features = [
     "ring",
     "hyper-rustls",
-] }
+], optional = true }
 moka = { version = "0.12.8", features = ["sync", "future"] }
 parse-size = { version = "1.1.0", features = ["std"] }
 prometheus = "0.14.0"
 prost = "0.13.3"
 prost-types = "0.13.3"
 rand = "0.8.5"
-rcgen = "0.13.1"
+rcgen = { version = "0.13.1", optional = true }
 reqwest = { version = "0.12.12", default-features = false, features = [
     "http2",
     "rustls-tls",
@@ -76,7 +78,7 @@ rustls = { version = "0.23.18", default-features = false, features = [
 rustls-acme = { version = "0.13.0", default-features = false, features = [
     "tls12",
     "ring",
-] }
+], optional = true }
 rustls-pemfile = "2.2.0"
 rustls-platform-verifier = "0.5.0"
 scopeguard = "1.2.0"
